+++ Controller/ProcessController.php
@@ -18,6 +18,7 @@ use Kontrolgruppen\CoreBundle\Entity\Process;
 use Kontrolgruppen\CoreBundle\Entity\ProcessLogEntry;
 use Kontrolgruppen\CoreBundle\Filter\ProcessFilterType;
 use Kontrolgruppen\CoreBundle\Form\ProcessCompleteType;
+use Kontrolgruppen\CoreBundle\Form\ProcessResumeType;
 use Kontrolgruppen\CoreBundle\Form\ProcessType;
 use Kontrolgruppen\CoreBundle\Repository\ProcessRepository;
 use Kontrolgruppen\CoreBundle\Repository\ProcessStatusRepository;
@@ -356,7 +357,7 @@ class ProcessController extends BaseController
      * @throws \Doctrine\ORM\NoResultException
      * @throws \Doctrine\ORM\NonUniqueResultException
      */
-    public function complete(Request $request, Process $process, ServiceRepository $serviceRepository, ProcessStatusRepository $processStatusRepository, EconomyService $economyService): Response
+    public function complete(Request $request, Process $process, ServiceRepository $serviceRepository, ProcessStatusRepository $processStatusRepository, ProcessManager $processManager): Response
     {
         $this->denyAccessUnlessGranted('edit', $process);
 
@@ -387,24 +388,9 @@ class ProcessController extends BaseController
                 $em->persist($lockedNetValue);
             }
 
-            $completedAt = new \DateTime();
-
-            $process->setCompletedAt($completedAt);
-            $process->setLastCompletedAt($completedAt);
-
-            $calculatedRevenue = $economyService->calculateRevenue($process);
-            $netCollectiveSum = $calculatedRevenue['netCollectiveSum'] ?: null;
-
-            if (!empty($process->getLastNetCollectiveSum())) {
-                $netCollectiveSumDifference = $netCollectiveSum - $process->getLastNetCollectiveSum();
-                $process->setNetCollectiveSumDifference($netCollectiveSumDifference);
-            }
-
-            $process->setLastNetCollectiveSum($netCollectiveSum);
-
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($process);
             $em->flush();
+
+            $processManager->completeProcess($process);
 
             return $this->redirectToRoute(
                 'process_show',
@@ -428,24 +414,44 @@ class ProcessController extends BaseController
     }
 
     /**
-     * @Route("/{id}/resume", name="process_resume", methods={"POST"})
+     * @Route("/{id}/resume", name="process_resume", methods={"POST", "GET"})
      *
      * @param \Symfony\Component\HttpFoundation\Request $request
      * @param \Kontrolgruppen\CoreBundle\Entity\Process $process
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function resume(Request $request, Process $process): Response
+    public function resume(Request $request, Process $process, ProcessStatusRepository $processStatusRepository): Response
     {
         $this->denyAccessUnlessGranted('edit', $process);
 
-        $process->setCompletedAt(null);
-        $process->setLockedNetValue(null);
-        $process->setLastReopened(new \DateTime());
-        $em = $this->getDoctrine()->getManager();
-        $em->persist($process);
-        $em->flush();
+        $form = $this->createForm(ProcessResumeType::class, $process, [
+            'available_statuses' => $processStatusRepository->getAvailableForProcess($process),
+        ]);
 
-        return $this->redirectToRoute('process_show', ['id' => $process->getId()]);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $process->setCompletedAt(null);
+            $process->setLockedNetValue(null);
+            $process->setLastReopened(new \DateTime());
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($process);
+            $em->flush();
+
+            return $this->redirectToRoute('process_show', ['id' => $process->getId()]);
+        }
+
+        return $this->render(
+            '@KontrolgruppenCore/process/resume.html.twig',
+            [
+                'menuItems' => $this->menuService->getProcessMenu(
+                    $request->getPathInfo(),
+                    $process
+                ),
+                'process' => $process,
+                'form' => $form->createView(),
+            ]
+        );
     }
 }