+++ Security/ProcessEconomyVoter.php
@@ -70,7 +70,7 @@ class ProcessEconomyVoter extends Voter
         }
 
         // ADMIN can do anything! The power!
-        if ($this->security->isGranted(['ROLE_SUPER_ADMIN', 'ROLE_ADMIN'])) {
+        if ($this->security->isGranted('ROLE_ADMIN')) {
             return true;
         }