+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.11.0] - 2021-02-23
 ### Fixed
 - [KON-404](https://jira.itkdev.dk/browse/KON-404): Provenu duplicates.
 
 ### Added
 - [SUPPORT-93](https://jira.itkdev.dk/browse/SUPP0RT-93): Added process delete console command.
+- [SUPPORT-101](https://jira.itkdev.dk/browse/SUPP0RT-101): Adding page for resuming processes
+- [SUPPORT-118](https://jira.itkdev.dk/browse/SUPP0RT-118): Enabling cache for PHPSpreadsheet
 
 ## [1.10.0] - 2021-02-15
 ### Added